Add new icons and create tab layout.
