+++ iOS-GreenRecipes/Storyboard/Base.lproj/Main.storyboard
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="18122" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="ag7-J6-tvY">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="18122" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="zrj-Gh-80g">
     <device id="retina6_1" orientation="portrait" appearance="light"/>
     <dependencies>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="18093"/>
@@ -8,10 +8,29 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--Navigation Controller-->
+        <!--Tab Bar Controller-->
+        <scene sceneID="tgY-L0-91c">
+            <objects>
+                <tabBarController id="zrj-Gh-80g" sceneMemberID="viewController">
+                    <tabBar key="tabBar" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" id="Wap-tf-hNe">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="49"/>
+                        <autoresizingMask key="autoresizingMask"/>
+                        <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                    </tabBar>
+                    <connections>
+                        <segue destination="ag7-J6-tvY" kind="relationship" relationship="viewControllers" id="Hha-oK-03H"/>
+                        <segue destination="ftH-Mx-IWo" kind="relationship" relationship="viewControllers" id="YTR-4d-Fdt"/>
+                    </connections>
+                </tabBarController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="grU-oS-W4u" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="-2906" y="349"/>
+        </scene>
+        <!--Home-->
         <scene sceneID="M0M-qC-NiQ">
             <objects>
                 <navigationController id="ag7-J6-tvY" sceneMemberID="viewController">
+                    <tabBarItem key="tabBarItem" title="Home" image="home" id="NIS-ke-qHy"/>
                     <navigationBar key="navigationBar" contentMode="scaleToFill" id="3ip-YI-oNJ">
                         <rect key="frame" x="0.0" y="44" width="414" height="44"/>
                         <autoresizingMask key="autoresizingMask"/>
@@ -22,7 +41,7 @@
                 </navigationController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="ffx-ut-PFf" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="-981" y="62"/>
+            <point key="canvasLocation" x="-2013" y="-108"/>
         </scene>
         <!--Recipes Home View Controller-->
         <scene sceneID="tne-QT-ifu">
@@ -38,7 +57,40 @@
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="-207" y="62"/>
+            <point key="canvasLocation" x="-1158" y="-108"/>
+        </scene>
+        <!--Search Recipes-->
+        <scene sceneID="1Ek-xJ-Zbj">
+            <objects>
+                <navigationController id="ftH-Mx-IWo" sceneMemberID="viewController">
+                    <tabBarItem key="tabBarItem" title="Search Recipes" image="recipe" id="MFZ-2a-wJN"/>
+                    <navigationBar key="navigationBar" contentMode="scaleToFill" id="4MW-VW-u5w">
+                        <rect key="frame" x="0.0" y="44" width="414" height="44"/>
+                        <autoresizingMask key="autoresizingMask"/>
+                    </navigationBar>
+                    <connections>
+                        <segue destination="SDl-x7-wNS" kind="relationship" relationship="rootViewController" id="zSB-a4-Gai"/>
+                    </connections>
+                </navigationController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="zNY-Q8-seO" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="-2025" y="719"/>
+        </scene>
+        <!--Recipes Search View Controller-->
+        <scene sceneID="cp8-5F-S6S">
+            <objects>
+                <viewController id="SDl-x7-wNS" customClass="RecipesSearchViewController" customModule="iOS_GreenRecipes" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="QM4-JH-kSB">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <viewLayoutGuide key="safeArea" id="X53-hD-qEe"/>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                    </view>
+                    <navigationItem key="navigationItem" id="PrM-ZA-VXX"/>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="10X-gy-dbl" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="-1183" y="719"/>
         </scene>
         <!--Recipe Detail View Controller-->
         <scene sceneID="ek8-9l-ht3">
@@ -53,10 +105,12 @@
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="ACf-r1-mFV" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="610" y="62"/>
+            <point key="canvasLocation" x="-297" y="909"/>
         </scene>
     </scenes>
     <resources>
+        <image name="home" width="48" height="48"/>
+        <image name="recipe" width="50" height="50"/>
         <systemColor name="systemBackgroundColor">
             <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
         </systemColor>
