+++ iOS-GreenRecipes/View Controllers/RecipesSearchViewController/RecipesSearchViewController.swift
@@ -1,0 +1,17 @@
+//
+//  RecipesSearchViewController.swift
+//  iOS-GreenRecipes
+//
+//  Created by Anupam Beri on 22/05/2021.
+//
+
+import UIKit
+
+class RecipesSearchViewController: UIViewController {
+
+  override func viewDidLoad() {
+    super.viewDidLoad()
+
+    overrideUserInterfaceStyle = .dark
+  }
+}
